requirements.txt
mcp>=1.6.0
anthropic>=0.50.0
python-dotenv>=1.0.1
json5>=0.9.24
fastapi>=0.115.0
uvicorn>=0.34.2
pydantic>=2.11.3